REST API works correctly - Implemented in Angular frontend
Upon application startup, the last 30 days worth of data is grabbed from
both datasets. This is now sent to both the Arrest-table-view and
Crime-table-view. You can change the values in the calendar as well as
the time, and press apply global filters to get new data back. The
tables update accordingly. Feel free to try and get the data working for
the other views and ask me any questions if you need to. BTW you can run
the backend and frontend separately, the frontend will call the backend
via proxy.