Latitudes and longitudes are added to the arrest database.
Arrests with "NO FIXED ADDRESS" are ignored: null values for lat/long cause the entries to be viewed as erroneous.

This code takes a few minutes to update the database because of all of the calls to Google's API.

We are also limited in how many calls to that API we can make daily, so try not to remake the table too many times.