Got rid of commented code, deleted unused import statements, and fixed indentation and other style issues.
Centered histogram in maximized view.

Made maximized tables shorter so that scrolling outside of the table is disallowed.

Map is not working in maximized view... code is commented out, but I'm not sure why it isn't working.

Pie chart looks bad in maximized view... Centering it doesn't work because the pie chart itself is left justified within the Google Charts component (I think).